requirements.txt
# Photo analysis (Claude vision)
anthropic>=0.49.0
# Search (TF-IDF, built-in — numpy only, no separate vector DB needed)
# numpy is installed as a transitive dep; add explicitly for clarity:
numpy>=1.22
# Optional OCR for scanned diary pages (also needs: brew install tesseract)
pytesseract>=0.3.10

annotated-types==0.7.0
anyio==4.12.1
certifi==2026.2.25
charset-normalizer==3.4.4
distro==1.9.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.13.0
openai==2.24.0
pillow==12.1.1
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.1
requests==2.32.5
sniffio==1.3.1
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
